revert back to int64 for observed generation. use patched json marshaller.

+++ pkg/certificates/agent/reconciliation/cert_agent_reconciler.go
@@ -116,12 +116,12 @@ func (r *certAgentReconciler) reconcileIssuedCertificate(
 	outputs certagent.Builder,
 ) error {
 	// if observed generation is out of sync, treat the issued certificate as Pending (spec has been modified)
-	if int64(issuedCertificate.Status.ObservedGeneration) != issuedCertificate.Generation {
+	if issuedCertificate.Status.ObservedGeneration != issuedCertificate.Generation {
 		issuedCertificate.Status.State = v1alpha2.IssuedCertificateStatus_PENDING
 	}
 
 	// reset & update status
-	issuedCertificate.Status.ObservedGeneration = uint32(issuedCertificate.Generation)
+	issuedCertificate.Status.ObservedGeneration = issuedCertificate.Generation
 	issuedCertificate.Status.Error = ""
 
 	// state-machine style processor
