+++ pkg/meshctl/checks/networking_crd_check.go
@@ -99,7 +99,7 @@ func (c *networkingCrdCheck) checkFailoverServices(fsList *v1alpha2.FailoverServ
 
 // This struct captures metadata common to all networking CRD statuses
 type networkingCrdStatus interface {
-	GetObservedGeneration() uint32
+	GetObservedGeneration() int64
 	GetState() v1alpha2.ApprovalState
 }
 
@@ -109,7 +109,7 @@ func (c *networkingCrdCheck) checkStatus(kind string, id ezkube.ResourceId, gene
 	if status.GetState() != v1alpha2.ApprovalState_ACCEPTED {
 		errorStrings = append(errorStrings, fmt.Sprintf("approval state is %s", status.GetState().String()))
 	}
-	if int64(status.GetObservedGeneration()) != generation {
+	if status.GetObservedGeneration() != generation {
 		errorStrings = append(errorStrings, fmt.Sprintf("observed generation does not match generation"))
 	}
 	if len(errorStrings) > 0 {
